generalized the checkout from git repositories
Instead of the function specialGaudiCheckout we use gitCheckout, which
requires a new (optional) field in the project description:
checkout_opts. This is a generic dictionary that can be used to pass
special arguments to a checkout function, that in the case of
gitCheckout are the url (mandatory) and the commit (optional).