generate build logs compatible with old ones (for CMake)
The HTML files produced have the same look and feel of the old ones,
with some improvements.

The main difference is the absence of the checkout log, because it
happens in a different Jenkins job (it can be added with some changes
in the jobs configurations).

The environment log has to be redefined (for the moment is a plain dump
of the system environment).

Since CMake does not group the build in packages, there is not such
division in the log report.

Support for CMT builds is not yet in place.

LBCORE-8
https://its.cern.ch/jira/browse/LBCORE-8