separation of artifacts location from build and deploy with rsync
In order to publish the artifacts to a shared location, the concept of
an 'artifacts directory' has been introduced (instead of 'sources' and
'build') and a special option to declare the deployment destination (via
rsync) has been added. The deployment is called (in background) after
each project is build, and once ad the end of the whole build (for
safety).

The new option --rsync-dest is parallel and complementary to
--deploy-reports-to.