added support for list of platforms in the slot configuration
For backward compatibility and to simplify the configuration, the list
of platforms to build can be specified in the configuration file as
'default_platforms' (with conversion from XML).
A script to print the list of platforms has been added with its Jenkins
wrapper that creates a property file in the artifacts directory.