fixed LBCORE-516: Added 'release' CouchApp.
The release directory contains the files for the couchapp

The application was developed based on the dashboard couchapp.

Feature 1: The Rebuild button, allows the user to send a request
	   to JENKINS and rebuild the slot with the given projects
	   and platforms.
Feature 2: After the rebuild is clicked, the confirmation dialog
	   appears with related information and the user can either
	   process or cancel.
Feature 3: The slots are not separated by date, they are grouped
	   to show only tha last weeks release-builds.
Feature 4: The applying filtering rules can only be projects.
Feature 5: Each slot has a link(direct link) to a url query that
	   represents only that slot.
Feature 6: When in queried page, the user has the "View All" option
	   to make all the slots appear in the page.

https://its.cern.ch/jira/browse/LBCORE-516